Pantalla notas guardadas modificada e integrada en UserScreen, listo para añadir funcionalidad de notas propias y notas con like
