+++ app/src/main/java/com/example/client/data/remote/service/NoteService.kt
@@ -1,7 +1,6 @@
 package com.example.client.data.remote.service
 
 import com.example.client.data.model.NoteDTO
-import com.example.client.data.model.NoteMapDTO
 import com.example.client.domain.model.note.NoteType
 import retrofit2.Response
 import retrofit2.http.Body
