+++ app/src/main/java/com/example/client/data/repositories/NoteRepository.kt
@@ -4,8 +4,8 @@ import com.example.client.common.NetworkResult
 import com.example.client.data.model.NoteDTO
 import com.example.client.data.model.NoteMapDTO
 import com.example.client.data.remote.datasource.NoteRemoteDataSource
-import com.example.client.domain.model.note.NoteType
 import com.example.client.di.IoDispatcher
+import com.example.client.domain.model.note.NoteType
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
