+++ app/src/main/java/com/example/client/domain/usecases/note/GetGroupedNotesByZoomUseCase.kt
@@ -1,5 +1,2 @@
 package com.example.client.domain.usecases.note
 
-import com.example.client.data.repositories.NoteRepository
-import javax.inject.Inject
-
