+++ app/src/main/java/com/example/client/ui/navigation/Navigation.kt
@@ -1,5 +1,7 @@
 package com.example.client.ui.navigation
 
+import androidx.compose.animation.EnterTransition
+import androidx.compose.animation.ExitTransition
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Add
@@ -16,6 +18,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.dimensionResource
+import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
@@ -23,18 +26,14 @@ import androidx.navigation.compose.rememberNavController
 import androidx.navigation.toRoute
 import com.example.client.R
 import com.example.client.ui.common.TopBar
-import com.example.client.ui.noteMap.list.NoteMapScreen
 import com.example.client.ui.normalNoteScreen.detail.NoteDetailScreen
 import com.example.client.ui.normalNoteScreen.list.NoteListScreen
-import com.example.client.ui.savedNotes.SavedScreen
-import com.example.musicapprest.ui.common.BottomBar
-import kotlinx.coroutines.launch
-import androidx.compose.animation.EnterTransition
-import androidx.compose.animation.ExitTransition
-import androidx.hilt.navigation.compose.hiltViewModel
+import com.example.client.ui.noteMap.list.NoteMapScreen
 import com.example.client.ui.noteMap.search.MapSearchScreen
 import com.example.client.ui.noteMap.search.SharedLocationViewModel
 import com.example.client.ui.userScreen.detail.UserScreen
+import com.example.musicapprest.ui.common.BottomBar
+import kotlinx.coroutines.launch
 
 @Composable
 fun Navigation() {
@@ -104,11 +103,16 @@ fun Navigation() {
 
         ) {
             composable<NormalNoteListDestination> {
-                NoteListScreen(showSnackbar = { showSnackbar(it) }, onNavigateToDetail = {navController.navigate(NormalNoteDetailDestination(it))})
+                NoteListScreen(
+                    showSnackbar = { showSnackbar(it) },
+                    onNavigateToDetail = { navController.navigate(NormalNoteDetailDestination(it)) })
             }
             composable<NormalNoteDetailDestination> { backStackEntry ->
                 val destination = backStackEntry.toRoute() as NormalNoteDetailDestination
-                NoteDetailScreen(noteId = destination.noteId, showSnackbar = { showSnackbar(it) }, onNavigateBack = { navController.navigateUp() })
+                NoteDetailScreen(
+                    noteId = destination.noteId,
+                    showSnackbar = { showSnackbar(it) },
+                    onNavigateBack = { navController.navigateUp() })
             }
             composable<NoteMapDestination> {
                 NoteMapScreen(
@@ -116,11 +120,6 @@ fun Navigation() {
                     onNavigateToList = { navController.navigate(MapSearchDestination) },
                     sharedLocationViewModel = sharedLocationViewModel
                 )
-            }
-
-
-            composable<NoteSavedListDestination> {
-                SavedScreen(showSnackbar = { showSnackbar(it) })
             }
 
             composable<MapSearchDestination> {
@@ -140,12 +139,6 @@ fun Navigation() {
                     showSnackbar = { showSnackbar(it) }
                 )
             }
-
-
-
-
-
-
 
         }
     }
