+++ app/src/main/java/com/example/client/ui/noteMap/list/NoteMapState.kt
@@ -2,7 +2,6 @@ package com.example.client.ui.noteMap.list
 
 import android.location.Location
 import com.example.client.data.model.NoteDTO
-import com.example.client.data.model.NoteMapDTO
 import com.example.client.domain.model.note.NoteType
 import com.example.client.ui.common.UiEvent
 import com.google.android.gms.maps.model.LatLng
