+++ app/src/main/java/com/example/client/ui/userScreen/detail/UserEvent.kt
@@ -3,4 +3,5 @@ package com.example.client.ui.userScreen.detail
 sealed class UserEvent {
     data object LoadUser: UserEvent()
     data object AvisoVisto : UserEvent()
+    data class SelectTab(val tab: UserTab) : UserEvent()
 }
