+++ app/src/main/java/com/example/client/ui/userScreen/detail/UserState.kt
@@ -1,10 +1,15 @@
 package com.example.client.ui.userScreen.detail
 
+import com.example.client.data.model.NoteDTO
 import com.example.client.data.model.UserDTO
 import com.example.client.ui.common.UiEvent
 
 data class UserState(
     val isLoading: Boolean = false,
     val aviso: UiEvent? = null,
-    val user: UserDTO? = null
+    val user: UserDTO = UserDTO(),
+    val selectedTab: UserTab = UserTab.NOTES,
+    val notes: List<NoteDTO> = emptyList()
 )
+
+enum class UserTab { NOTES, FAVORITES, LIKES }
