+++ app/src/main/java/com/example/client/ui/userScreen/detail/UserViewModel.kt
@@ -3,18 +3,21 @@ package com.example.client.ui.userScreen.detail
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.client.common.NetworkResult
+import com.example.client.domain.usecases.social.GetNoteSavedUseCase
 import com.example.client.domain.usecases.user.GetUserUseCase
 import com.example.client.ui.common.UiEvent
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
 class UserViewModel @Inject constructor(
-    private val getUserUseCase: GetUserUseCase
+    private val getUserUseCase: GetUserUseCase,
+    private val getNoteSavedUseCase: GetNoteSavedUseCase,
 ) : ViewModel() {
 
     private val _uiState = MutableStateFlow(UserState())
@@ -24,6 +27,16 @@ class UserViewModel @Inject constructor(
         when (event) {
             is UserEvent.LoadUser -> loadUser()
             is UserEvent.AvisoVisto -> avisoVisto()
+            is UserEvent.SelectTab -> {
+                _uiState.value = _uiState.value.copy(selectedTab = event.tab)
+                if (event.tab== UserTab.FAVORITES) {
+                    getSavedNotes()
+
+                }else {
+                    _uiState.value = _uiState.value.copy(notes = emptyList())
+                }
+            }
+                
         }
     }
 
@@ -34,17 +47,53 @@ class UserViewModel @Inject constructor(
                 is NetworkResult.Success -> {
                     _uiState.value = _uiState.value.copy(
                         isLoading = false,
-                        user = result.data
+                        user = result.data,
+                        selectedTab = UserTab.NOTES
                     )
                 }
                 is NetworkResult.Error -> {
                     _uiState.value = _uiState.value.copy(
                         isLoading = false,
-                        aviso = UiEvent.ShowSnackbar(result.message)
+                        aviso = UiEvent.ShowSnackbar(result.message),
+                        selectedTab = UserTab.NOTES
                     )
                 }
                 is NetworkResult.Loading -> {
                     _uiState.value = _uiState.value.copy(isLoading = true)
+                }
+            }
+        }
+    }
+
+    private fun getSavedNotes() {
+        viewModelScope.launch {
+            _uiState.update { it.copy(isLoading = true) }
+
+            when (val result = getNoteSavedUseCase()) {
+                is NetworkResult.Success -> {
+                    _uiState.update {
+                        it.copy(
+                            notes = result.data,
+                            isLoading = false
+                        )
+                    }
+                }
+
+                is NetworkResult.Error -> {
+                    _uiState.update {
+                        it.copy(
+                            aviso = UiEvent.ShowSnackbar(result.message),
+                            isLoading = false
+                        )
+                    }
+                }
+
+                is NetworkResult.Loading -> {
+                    _uiState.update {
+                        it.copy(
+                            isLoading = true
+                        )
+                    }
                 }
             }
         }
